requirements.txt
openai
playwright
python-dotenv